adc: change pins to their new locations
We were getting some sort of interference: the LoRa module was unable to communicate properly when the ADC's Dout line was connected (even with the ADC disabled).
Adding a filter capacitor between Dout and GND helped the ADC give zero consistently when disabled, but the LoRa module still wouldn't work.

Adding a cap near the LoRa's MISO pin and/or near the RPi's might help, but I don't have the components to test.